pyproject.toml
[tool.poetry]
name = "workers-tree"
version = "0.1.0"
description = ""
authors = ["Dddarknight <9801677@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<4"
python-dotenv = "^0.21.0"
dj-database-url = "^1.0.0"
Django = "^4.1.2"
django-bootstrap4 = "^22.2"
django-crispy-forms = "^1.14.0"
django-tables2 = "^2.4.1"
django-filter = "^22.1"
django-htmx = "^1.12.2"
crispy-bootstrap5 = "^0.7"
Pillow = "^9.3.0"

[tool.poetry.dev-dependencies]
coverage = "^6.5.0"
flake8 = "^5.0.4"
pytest = "^7.2.0"
django-debug-toolbar = "^3.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
